Added the other sensor and the user image
I also made another version that uses skeleton tracker to track the hand
and then looks for the closest point within that region and applies the
flood fill on top. It doesn’t jump to the corners any more. I haven’t
pushed that one but it’s under the same github directory named
CSE481-3DAuth-siavash