Added Face tracking start/stop, and password point retrieval from the xml file
Use FaceTrackingViewer.StopTracking() and
FaceTrackingViewer.StartTracking() when you need to start/stop the
tracker. No need to dispose or register a null sensor. Also, if an
already registered user got his face recognizer, the User object is
populated with that users password points.
I updated users.xml to support this. So make sure that the xml in the
bin/debug folder is in the same format as the one in this commit.
We'll figure out the information flow for new user registration
tomorrow.